[NT-0] VisionOS Support for CSP (#351)
[NT-0] feat: VisionOS support for CSP.

Notable changes include:

* A new shell script for generating VisionOS xcodeprojects: `generate_solution_visionos.sh`. Run this, open the project and build to create visionos-compatible CSP libraries.
* Since premake does not yet support the device, the approach taken is to instruct premake to build for iOS, with an additional argument flagging that it should actually generate a project targeting visionos. Responding to this argument is primarily handled within the `premake5_helpers.lua` script.
* Some refactorings have been made around the Unity package generation pipeline to allow for Unity-compatible packages to be generated via MacOS. This greatly helps with local debugging and iteration.
* Mimalloc has been disabled for iOS and VisionOS. Whilst this has not been reported to be an issue on iOS, it causes problems on VIsionOS similar to what we have seen in the past for Android and WASM. I would recommend we strip it out entirely in the near future.
* A fix for a MacOS bug in our project generation scripts which always led to a loose file named '142' being created at the root directory level.
