+++ generate_solution_ios.sh
@@ -16,18 +16,12 @@ fi
 
 if [ ! -f /modules/premake/bin/release/premake5 ]
 then
-	if [ $((${XCODE_VERSION//.})) > 142 ]; then
-		mkdir -p ./modules/premake/bin/release/
-		cd ./modules/premake/bin/release/
-		curl -O -L https://github.com/premake/premake-core/releases/download/v5.0.0-beta2/premake-5.0.0-beta2-macosx.tar.gz
-		tar -xzvf premake-5.0.0-beta2-macosx.tar.gz
-		rm premake-5.0.0-beta2-macosx.tar.gz
-		cd ../../../..
-	else
-		cd ./modules/premake/
-		make -f Bootstrap.mak macosx
-		cd ../..
-	fi
+    mkdir -p ./modules/premake/bin/release/
+    cd ./modules/premake/bin/release/
+    curl -O -L https://github.com/premake/premake-core/releases/download/v5.0.0-beta2/premake-5.0.0-beta2-macosx.tar.gz
+    tar -xzvf premake-5.0.0-beta2-macosx.tar.gz
+    rm premake-5.0.0-beta2-macosx.tar.gz
+    cd ../../../..
 fi
 
 ./modules/premake/bin/release/premake5 --os=ios xcode4 "$@"
